+++ src/ResourceManager/Compute/Commands.Compute/VirtualMachine/Operation/NewAzureVMCommand.cs
@@ -296,8 +296,11 @@ namespace Microsoft.Azure.Commands.Compute
                     name: _cmdlet.SecurityGroupName,
                     openPorts: _cmdlet.OpenPorts);
 
+                bool enableAcceleratedNetwork = Utils.DoesConfigSupportAcceleratedNetwork(_client,
+                    ImageAndOsType, _cmdlet.Size, Location, DefaultLocation);
+
                 var networkInterface = resourceGroup.CreateNetworkInterfaceConfig(
-                    _cmdlet.Name, subnet, publicIpAddress, networkSecurityGroup);
+                    _cmdlet.Name, subnet, publicIpAddress, networkSecurityGroup, enableAcceleratedNetwork);
 
                 var availabilitySet = _cmdlet.AvailabilitySetName == null
                     ? null